Use usize coordinates to index buffers without numeric conversion

+++ src/theme.rs
@@ -1,4 +1,4 @@
-pub const GUTTER_WIDTH: u16 = 5;
+pub const GUTTER_WIDTH: usize = 5;
 
 // pub const GUTTER_DEFAULT: &[u8] = b"\x1b[38;2;100;100;100m\x1b[48;2;60;60;60m";
 // pub const GUTTER_HIGHLIGHT: &[u8] = b"\x1b[38;2;200;200;200m\x1b[48;2;60;60;60m";
